+++ projects/ngx-uploader-directive/src/public-api.ts
@@ -6,5 +6,5 @@ export * from './lib/configs/config';
 export * from './lib/models/ngx-uploader-directive-models';
 export * from './lib/directives/ngx-uploader-drop.directive';
 export * from './lib/directives/ngx-uploader-select.directive';
-export * from './lib/ngx-uploader-directive.service';
+export * from './lib/services/ngx-uploader-directive.service';
 export * from './lib/ngx-uploader-directive.module';